total search 인코딩 테스트
